moved atom
