+++ src/components/atoms/index.tsx
@@ -1,4 +1,3 @@
 export * from './input'
 export * from './textarea'
 export * from './button'
-export * from './echoButton'
